src/requirements.txt
aiofiles==25.1.0
pandas==2.2.3
fastapi==0.119.1
fastapi-health==0.4.0
Flask==3.1.2
flask-cors==6.0.1
fireworks-ai==0.15.12
langgraph==0.6.6
prometheus-client==0.23.1
python-dotenv==1.0.0
beautifulsoup4==4.12.2
uuid==1.30
sqlalchemy==2.0.38
starlette-exporter==0.23.0
psycopg2-binary==2.9.10
pydantic-settings==2.11.0
pyarrow==19.0.1
pydub==0.25.1
librosa==0.11.0
uvicorn==0.34.3
python-multipart==0.0.20
passlib==1.7.4
bcrypt==3.2.0
celery==5.5.3
redis==7.0.0
kombu==5.5.4
billiard==4.2.2
vine==5.1.0
flower==2.0.1